# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management System",
    "summary": "Multi-tenant school management platform",
    "description": "EduCloud API -- identity, tenancy, RBAC, plans and invitations.\n\n**Tenant isolation** is enforced by PostgreSQL Row-Level Security. The `org` claim in your access token scopes every query at the database level; cross-tenant reads return 404, never 403, so resource existence is never disclosed.\n\n**School scoping** is a second, softer boundary inside the organization: a school-scoped membership sees only its own campus, while an organization owner spans all of them.\n\n**Permissions** are `resource:action` strings resolved per request from the caller's role. They are deliberately not embedded in the token, so revoking access takes effect on the very next request.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Me",
        "description": "The caller's identity, contexts and resolved permissions (spec \u00a78).\n\nOne call, because the app shell needs all of it before it can render anything:\nthe nav depends on permissions, the header on the active school, the switcher on\nthe membership list.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
    "/api/v1/schools/{school_id}/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_v1_schools__school_id__members_get",
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "School Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberRead"
                  },
                  "title": "Response List Members Api V1 Schools  School Id  Members Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MembershipSummary"
            },
            "type": "array",
            "title": "Memberships"
          },
          "active_membership_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Membership Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "organization_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Status"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "school_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Name"
          },
          "role_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Code"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "full_name",
          "avatar_url",
          "locale",
          "memberships",
          "active_membership_id",
          "organization_id",
          "organization_name",
          "organization_status",
          "school_id",
          "school_name",
          "role_code",
          "permissions"
        ],
        "title": "MeResponse",
        "description": "`GET /auth/me` -- everything the app shell needs on boot.\n\nReturns the resolved permission set so the frontend can hide actions the user\ncannot perform. That hiding is COSMETIC ONLY (spec \u00a79): every action is\nre-checked server-side by `require()`. A frontend treating this list as the\nsecurity boundary is one `curl` away from being bypassed."
      },
      "MemberRead": {
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid",
            "title": "Membership Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "title": "Role Id"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "joined_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Joined At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "membership_id",
          "user_id",
          "email",
          "full_name",
          "school_id",
          "role_id",
          "role_code",
          "role_name",
          "status",
          "is_primary",
          "joined_at",
          "created_at"
        ],
        "title": "MemberRead"
      },
...
      "MembershipSummary": {
        "properties": {
          "membership_id": {
            "type": "string",
            "format": "uuid",
            "title": "Membership Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "organization_name": {
            "type": "string",
            "title": "Organization Name"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "school_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Name"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "is_org_level": {
            "type": "boolean",
            "title": "Is Org Level"
          }
        },
        "type": "object",
        "required": [
          "membership_id",
          "organization_id",
          "organization_name",
          "school_id",
          "school_name",
          "role_code",
          "role_name",
          "is_primary",
          "is_org_level"
        ],
        "title": "MembershipSummary",
        "description": "One context the user may act in. Rendered by the context switcher."
      },
...
    }
  }
}